OKAY, this is alot.... 1. Created Skeleton Classes for basic; Run program, Menu display, create user, load wallet, record transaction, etc... however far we want to take it. I included as many TODO's as possible, but there are alot more. WE NEED to do more in depth Comments, with ATTRIBUTES and REPRESENTATIONAL INVARRIENTS so all the variables are right when we finish the coding. I created a basic outline for wallettest so we can create funds testing as we go. I will also create tests for user creation soo (Student) but I am still trying to figure out the format to record user info so that we can polymorphically access it in the commands.
THIS IS THE STRUCTURE I HAVE CREATED>>> IF You guys want to change it let me know!!!
The Shell that runs it all is tCardAppShell.java it maintains the run state, read its description. It instances the Menu class. The Menu class will be the user interface sending information to and from the command line, and it will send information to UserCommands.java this is the 'Business Rules' or User case level of operations where we abstract the menu requests into actual functionality that interacts with user and wallet entities, it also serves as a defacto data base(still need to think about how to hard store info in a csv file or text)

Soo, please let me know thoughts on the structure. I want to keep it simple then we can extend and refactor to change... for instance i think the wallet class might need to be a child of a card class if we consider that Tcard transactions aren't always financial (ie door, library, etc) but we could still give those functionality by keeping a record of them.

I haven't created any web/virtual web interfaces yet. For now, I thought it might be easier to load student info as a "previous Load State" as a temporary workaround... again, anybody that knows how to do this should step up and take the lead.

If it looks like we are on the right track? Then I need help with a number of things. Interfacing with the command line would be a great task in the Menu class todo's. Someone would just need to create basic output and input retrieval. The string contents for the menu will be a mix of stored strings and formated class info... Which would also be another good task for someone, create the format strings for each class.

Please post feedback and notes both on GIT (so we have a record of all our teamwork!) and on the Discord project chat. Thanks Pete J

OH... Also, if this looks ok, can someone fulfill a merge pull request, as I should not since I wrote it.