Added CRC Model.
I tried to write out all of these with the Clean Architecture in mind, but it is possible that what I wrote for "Level" in a given card may be incorrect. There were a few cards that I didn't complete because I felt I didn't have the best understanding. I left TODOs in each of the CRC cards that I didnt finish, but for the sake of convenience, I will also list them here. The cards that still need to be finished are:
- Controller
- InputBoundary
- Menu
- tCardAppShell
- UserCommands