Added use case class that checks if a given user is allowed to access a given facility, based on the facilities criteria and the users information. A facilitys criteria is stored as a string in the following format "program=(program1/program2/...),level=(level1/level2/...):department=(department1/department2/...),position=(position1/position2/...)". Here are some details about the formatting - The left side of the ":" specifies the criteria for students, the right side of the ":" specifies the criteria for students. - The "/" represent OR and the "," represents AND. - If a certain field lists "any" (for example "program=(any)"), that means any possible option for that field is allowed. - If a certain field lists "none" (for example "program=(none)"), that means no one that the field applies to is allowed.
For example, if the user criteria for facility is "program=(math/physics/engineering),level=(undergrad):department=(compsci),position=(any)", that means
- undergraduate students that are in any of the math OR physics OR engineering programs are allowed to enter.
- ANY faculty that are in the department of compsci are allowed to enter.

I provided some sample usage in a commented out main method, and also gave comments throughout the evaluate method to make it easier to understand what is going on.

As always, I dont feel that this is necessarily the best way to do things and changes are of course still possible, so please feel free to comment or make suggestions/changes.