Updated the User, Faculty, and Student classes. Here are some notes about the changes:
In all of their constructors, I assumed that the parameter inputted is a string array that looks like: {first name, last name, UTORid, password, status, ID number, email, program, level, year, department, position}. This ordering corresponds to the ordering of the columns of the user database. So, when we instantiate a user, all we must do is put their respective row from the userdata base into an ArrayList, then feed it to the constructor. Note that regardless of the type of user, there will be columns for:
 - program, level, and year (attributes that are specific to students), and
 - department and position (attributes that are specific to faculty members)
This is to keep the rows of the database consistent. I assume if a users status is "student" (for example), their entries for department and position in the database will be "N/A".

The indicies for attributes in the user database will now not align with the indices in a users "profile" attribute (e.g. in the database, "email" will be the 7th column, but it will be the 6th element of the "profile" attribute). So, the attributes "FIRST_NAME", "LAST_NAME", "STATUS", etc, for a user will correspond to the indices of "profile", NOT the indices of the database columns.

In User, I made it so that a user's password is not added to their profile attribute. Instead, their password is written to a private attribute called "password". I also added a method to check if a password attempt is correct and a method to change the user's password. Both of these methods will be used when a user wants to change their password. We just have to make sure to update their password in the database after they change it themselves.