Refactored the user database accessor class UserDBHelper.
While refactoring the activity classes, I will be making use case classes that use methods in this class. However, since this class is technically a Gateway (it belongs in the interface adapters layer), use cases that depend on this class will violate clean architecture. My solution was to create a usecase-level interface UserReadWriter (which lists all the methods from UserDBHelper that the use cases will use) and make UserDBHelper implement this interface. This way, the use cases will depend only on UserReadWriter, which is in the use case level, so clean architecture is satisfied. I talked to Professor Calver about all of this, and we agreed that this was the correct way to refactor UserDBHelper in order to satisfy clean architecture.