Refactored RegisterActivity.
The Register Activity class was taking on too much responsibility because it was checking validity of user inputs by directly comparing against the database. Chris made it clear in the phase 1 feedback that these responsibilities should belong on the use case level. So, I made a controller class RegisterController that controls the user registration process and a use case class RegisterUseCase which is the place that takes the bulk of the responsibility of checking validity of user input. Because of these changes, the RegisterActivity class is shorter and everything related to it satisfies clean architecture.

Btw, I also did other minor refactoring such as changing format of variable names, changing format of method names, adjusting comments, etc.